testutil: move syscall checker to separate file
